separate cri and log streamer
