+++ logstreamer/pool.go
@@ -1,4 +1,4 @@
-package cri
+package logstreamer
 
 import (
 	"errors"
